* Solve conlfict in openspace.cfg and update use of format library

+++ src/navigation/orbitalnavigator_lua.inl
@@ -2,7 +2,7 @@
  *                                                                                       *
  * OpenSpace                                                                             *
  *                                                                                       *
- * Copyright (c) 2014-2023                                                               *
+ * Copyright (c) 2014-2024                                                               *
  *                                                                                       *
  * Permission is hereby granted, free of charge, to any person obtaining a copy of this  *
  * software and associated documentation files (the "Software"), to deal in the Software *
@@ -36,7 +36,7 @@ namespace {
 
     double is = node->interactionSphere();
     global::navigationHandler->orbitalNavigator().setMinimumAllowedDistance(
-        is * multiplier
+        static_cast<float>(is * multiplier)
     );
 }
 
@@ -52,7 +52,7 @@ namespace {
 
     double is = node->interactionSphere();
     global::navigationHandler->orbitalNavigator().setMaximumAllowedDistance(
-        is * multiplier
+        static_cast<float>(is * multiplier)
     );
 }
